Add empty state to command interface + make small design adjustments
