+++ packages/ui/src/components/domain/commandinterface/RecommendationRow.tsx
@@ -1,5 +1,18 @@
 import clsx from 'clsx'
 
+export type RecommendationRowType = {
+  Icon?: IconComponentType | null
+  menuText: string
+  pageRestriction: string | null
+  isIdPrimary?: boolean
+  subtitle?: string | null
+  isProjectPrimary?: boolean
+  commandInterfaceOptions: string
+  shortcutKey: string | null
+  needConfirmation?: boolean
+  hide?: boolean
+}
+
 type RecommendationRowProps = {
   Icon?: IconComponentType | null
   shortcutKey: string | null
